+++ api/src/Geography/Entity/Address.php
@@ -326,6 +326,7 @@ class Address implements \JsonSerializable
      * @var User|null The owner of the address.
      *
      * @ORM\ManyToOne(targetEntity="App\User\Entity\User", inversedBy="addresses")
+     * @ORM\JoinColumn(onDelete="CASCADE")
      */
     private $user;