+++ api/src/Migrations/Version20211012163300.php
@@ -1,0 +1,29 @@
+<?php declare(strict_types=1);
+
+namespace DoctrineMigrations;
+
+use Doctrine\DBAL\Schema\Schema;
+use Doctrine\Migrations\AbstractMigration;
+
+/**
+ * Auto-generated Migration: Please modify to your needs!
+ */
+final class Version20211012163300 extends AbstractMigration
+{
+    public function up(Schema $schema) : void
+    {
+        // this up() migration is auto-generated, please modify it to your needs
+        $this->abortIf($this->connection->getDatabasePlatform()->getName() !== 'mysql', 'Migration can only be executed safely on \'mysql\'.');
+
+        // Fix wrong sequence item
+
+        $this->addSql('UPDATE `gamification_action` SET `title` = \'Communauté rejointe\' WHERE `gamification_action`.`id` = 7');
+        $this->addSql('UPDATE `sequence_item` SET `gamification_action_id` = \'7\' WHERE `sequence_item`.`id` = 8');
+    }
+
+    public function down(Schema $schema) : void
+    {
+        // this down() migration is auto-generated, please modify it to your needs
+        $this->abortIf($this->connection->getDatabasePlatform()->getName() !== 'mysql', 'Migration can only be executed safely on \'mysql\'.');
+    }
+}