+++ api/src/Solidary/Entity/Subject.php
@@ -113,7 +113,7 @@ class Subject
      *
      * @Assert\NotBlank
      * @ORM\ManyToOne(targetEntity="App\Solidary\Entity\Structure", inversedBy="subjects")
-     * @ORM\JoinColumn(nullable=false)
+     * @ORM\JoinColumn(nullable=false, onDelete="CASCADE")
      * @Groups({"writeSolidary"})
      * @MaxDepth(1)
      */