+++ api/src/Solidary/Service/SolidaryUserManager.php
@@ -610,6 +610,11 @@ class SolidaryUserManager
         $solidaryUserStructure->setStructure($solidaryBeneficiaryStructure);
         $solidaryUserStructure->setSolidaryUser($solidaryUser);
 
+        //we check if the structure need proofs before validation if not we validate automaticaly the candidate
+        if (count($solidaryUserStructure->getStructure()->getStructureProofs()) == 0) {
+            $solidaryBeneficiary->setValidatedCandidate(true);
+        }
+
         if ($solidaryBeneficiary->isValidatedCandidate()) {
             // Already accepted. We set the date a give the appropriate role to the user
             $solidaryUserStructure->setAcceptedDate(new \Datetime());
@@ -799,6 +804,11 @@ class SolidaryUserManager
         $solidaryUserStructure->setStructure($solidaryVolunteerStructure);
         $solidaryUserStructure->setSolidaryUser($solidaryUser);
 
+        //we check if the structure need proofs before validation if not we validate automaticaly the candidate
+        if (count($solidaryUserStructure->getStructure()->getStructureProofs()) == 0) {
+            $solidaryVolunteer->setValidatedCandidate(true);
+        }
+        
         if ($solidaryVolunteer->isValidatedCandidate()) {
             // Already accepted. We set the date a give the appropriate role to the user
             $solidaryUserStructure->setAcceptedDate(new \Datetime());
@@ -807,6 +817,25 @@ class SolidaryUserManager
             $userAuthAssignment = new UserAuthAssignment();
             $userAuthAssignment->setAuthItem($authItem);
             $user->addUserAuthAssignment($userAuthAssignment);
+        }
+
+        // Proofs
+        if ($solidaryVolunteer->getProofs()) {
+            foreach ($solidaryVolunteer->getProofs() as $givenProof) {
+                // We get the structure proof and we create a proof to persist
+                $structureProofId = null;
+                if (strrpos($givenProof['id'], '/')) {
+                    $structureProofId = substr($givenProof['id'], strrpos($givenProof['id'], '/') + 1);
+                }
+                    
+                $structureProof = $this->structureProofRepository->find($structureProofId);
+                if (!is_null($structureProof) && isset($givenProof['value']) && !is_null($givenProof['value'])) {
+                    $proof = new Proof();
+                    $proof->setStructureProof($structureProof);
+                    $proof->setValue($givenProof['value']);
+                    $solidaryUserStructure->addProof($proof);
+                }
+            }
         }
 
         $solidaryUser->addSolidaryUserStructure($solidaryUserStructure);